Show, insert, update and delete artkworks on a list.
First commit should have been after completing the list (show part).
First of all create a Class Obra, with all the atributes listed on the gson file.
Created a Jlist in the design menu to copy the necessary things in order to create a Jlist of objects. Once all is copied, delete the Jlist and add a ScrollPanel, inside it declare a Jlist<Obra>.
After that create an ArrayList<Obra> where all the artworks from the gson will be introduced.
Import gson .jar, declare it and read the gson file introducing the data on the ArrayList.
After reading the gson file, update the Jlist listmodel to show them on the list.
Create a menu bar with menu CRUD and File, on CRUD add menu items: insert, update, delete; on File add menu items: save, load, exit.

Second commit should have been after completing insert dialog (without images yet).
First of all create a new Dialog, called InsertDialog.
Declare in its design all atributes that need to be inserted, txtFields for the atributes.
Create a boolean on the dialog, put it true when the insert button is clicked.
If the boolean is true, on the mainform, get all the fields from the dialog and write them on a new Obra object. Then add it to the ArrayList and update the ListModel.

Third commit should have been after completing update dialog (without images yet).
First of all create a new Dialog, called UpdateDialog.
Declare in its design all atributes that need to be updated, txtFields for the atributes.
Add all the txtFields and labels, except for the ones for the num registre, on a panel and hide it (setVisible(false)).
Add a button to show all the elements on the panel when clicked if the num Registre is on the arrayList of obra, also writes all the atributes of the obra on the txtFields.
Create a boolean on the dialog, put it true when the insert button is clicked.
If the boolean is true, on the mainform, get all the fields from the dialog and update them on the Obra object that was selected and update the ListModel.
Also create an event MouseClicker on the List, when clicked twice, get the num Registre of the object and send it to the Update Dialog, run an Update Dialog with the num Registre already writen.

Forth commit should have been after completing delete dialog.
First of all create a new Dialog, called DeleteDialog.
Create a combobox on the design to copy all necessary events to use it on a manual created object comboBox.
Then create a panel, inside which, the comboBox will be placed.
Create a ComboBox of Obra, insert the elements of they ArrayList of Obra into the ComboBoxModel.
Create a boolean in the dialog, set it true when the delete button is clicked.
If the boolean is true, on the mainform, remove from the ArrayList the element selected to delete and update the ListModel.

Fifth commit should have been after adding images.
First of all add an atribute imatge to the Obra class and another atribute imagePath, which is declared as transient, meaning that it will not be written on the gson file.
On the insert dialog and on the update dialog, add a textless label and a button (on the update dialog, both items should be inside the panel).
On the button, define a FileChooser, that will load the absolutePath into a method which sets the image on the label's icon property. Also calls a method resizeImage that resizes the image to fit the label's size.
Now in order to create obras and updating them, the image should be added, so the mainform insert and update should add that, getting the fileChooser().getSelectedFile().getName() and getParent() for imagePath.
Also add a noObra image on a package spdvi.image, in order to put that image when no image is selected.
Do a try-catch on the insert from mainform where asks if the filechooser is null, if its null sets the file with the image noObra.
On the update dialog, to set the image from the obra selected, the imagePath + image name is what sets the image, the selected file on fileChooser is also setted the same way. Creating a new File with the imagePath + image name as parameters.
Lastly on the load of gson file into the list, create a for:each of obras that sets the imagePath to users/.../AppData/Local/OpusList/Images.

Sixth and last commit, adding save button, correcting errors.
First of all add a buton save and a button load.
On the save button iterate the list to add the images into the folder imageFolder which is users/.../AppData/Local/OpusList/Images, add them with the Registre num and .jpg, to do that, set the image of every obra to be the registre num.
Then create a gson and a writer in order to write the data of the list into the json file located on users/.../AppData/Local/OpusList/obres.json.
The image on the json file then, its added as the registre num aswel.
After that create a new ArrayList of String in which will be placed the name of the images that are on the json.file in order to being able to delete the ones that are no longer on the file but are on the images directory, iterating the names of the images and comparing them to the ones on the arrayList.
Create a dialog to confirm save.
After that create a boolean dataChanged, when data is changed like insert, udpate, delete, set it to true and when data is saved or loaded, set it false.
Create a WindowsClosing event, if dataChanged is true, the confirmSave dialog will now show that there is data changed and it will do as the button save.
After that create a method loadOpus, where the code that was on the windowsOpened event will be placed. On the windowsOpened event and load button and load menu, will call this method.
Also the save menu will doClick on the save button, and on the exit menu will call the windowsClosing event.
Lastly manage errors from inserting empty data, or if there is an artwork with a register number equal to the one that is going to be inserted.